pyproject.toml
[project]
name = "cofee"
version = "1.0.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "celery>=5.6.0",
    "django>=6.0",
    "django-cors-headers>=4.9.0",
    "django-nested-inline>=0.4.6",
    "django-redis>=6.0.0",
    "django-storages[boto3]>=1.14.6",
    "django-uuid7>=0.1.1",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt>=5.5.1",
    "drf-yasg>=1.21.11",
    "msgpack>=1.1.2",
    "pillow>=12.0.0",
    "psycopg[binary,pool]>=3.3.1",
    "python-dotenv>=1.2.1",
    "pytz>=2025.2",
    "scikit-learn>=1.6.0",
]

[dependency-groups]
dev = [
    "django-debug-toolbar>=6.1.0",
    "ipdb>=0.13.13",
]
